nvme: ignore error messages when running tests
There are subtle differences in the way fc is creating associations,
causing the error messages from 'nvme connect' to be different than
on other transports. As we'll get notified anyway when 'nvme connect'
fails there is not point is specifying the error message in the response
files, so just suppress error messages here.

Signed-off-by: Hannes Reinecke <hare@suse.de>