block/{014,015,030}: don't require modular null_blk for fault-injection
This blktests change changes null_blk fault-injection settings to be
configured via configfs instead of module parameters.
This allows null_blk fault-injection tests to run even if the null_blk is
built-in the kernel and not built as a module.

If the null_blk does not yet support configuring fault-injection via
configfs, fall back to set up with module parameter.

Signed-off-by: Akinobu Mita <akinobu.mita@gmail.com>
[Shin'ichiro: improved commit title]
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>