nvme/{041,042,043,044,045,048}: Use default hostnqn and hostid
The host might have enabled the udev/systemd auto connect feature.
This disturbs the blktests for the fc transport. nvme-cli is able
to distinguish between the different invocations via the --context
option.

Instead creating random generated IDs or reuse the hostnqn, it's safer
to always use the same hostnqn to reduce the risk that the matching
logic doesn't work. For this purpose, drop references to
/etc/nvme/hostid and /etc/nvme/hostnqn.

Signed-off-by: Daniel Wagner <dwagner@suse.de>
[Shin'ichiro: resolved conflict and noted /etc/nvme/* in commit message]
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>