Revert "Do not suppress any shellcheck warnings"
This reverts commit 26664dff17b626550fae99181112f9ceff9bb971.

The commit enabled the shellcheck warning SC2119 based on the
understanding that bash would pass arguments list "$@" from caller to
callee when functions are called without arguments. However, it was not
correct then merit of SC2119 is not so important. On the other hand,
SC2119 reports false positive warning when bash functions take optional
arguments and they are called without arguments.

Per discussion on the list [1], we concluded that the merit of optional
arguments is larger than that of SC2119. Hence, disable SC2119 again.

[1] https://lore.kernel.org/linux-nvme/o5xnqvujzakhrudv7p64owiuzgozmean6blxow4vdxhdqozg5v@qznf2tzmey7k/

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>