nvme: Do not print UUID to log files
The UUID/wwid of a namespace might be assigned externally, so
we shouldn't register it in the 'out' files.
The current checks for UUID/wwid are just there to ensure that
if a UUID is present is should match the wwid settting.
So rather add a function _check_uuid() which does exactly that
and don't register the actual UUID in the 'out' files.

Signed-off-by: Hannes Reinecke <hare@suse.de>