common/nvme, nvme/{rc,054}: remove "local -n" from _get_nvmet_ports()
The -n option of the bash local variable declaration declares that
the variable is used as nameref. However, this nameref feature was
introduced to bash at its version 4.3. While blktests requires bash
version 4.2. To not rely on the bash version 4.3, replace the nameref
feature with a loop to parse arguments.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>