nvme: introduce wwnn/wwpn helper functions
Commit "nvme: sanitize transport parameter in _create_nvmet_port()"
modified the wwnn and wwpn names for fc transport to calculate from the
port ID. It also moved the timing to set the global variable
def_host_traddr based on the wwnn and wwpn from _setup_nvmet() to
_create_nvmet_port(). This move made _nvme_connect_subsys() fail
to refer to the def_host_traddr variable, and caused nvme/003 test case
failure.

To avoid the failure, do the wwnn and wwpn calculations both in
_create_nvmet_port() and _nvme_connect_subsys(). Introduce helper
functions to calculate wwnn and wwpn.

Reviewed-by: Daniel Wagner <dwagner@suse.de>
Reviewed-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>