nvme/{rc,002,016,017,030}: pass port ID to genctr helper functions
Commit "nvme: sanitize transport parameter in _create_nvmet_port()"
introduced wwnn and wwpn calculation based on port ID for fc transport.
This means that --traddr and --host-traddr options for "nvme discover"
command should reflect the port ID. This change was missing for the
genctr, or Generation Counter, related helper functions and it made
the test case nvme/030 fail.

To avoid the failure, pass the port ID to the genctr related helper
functions and reflect the port ID for fc transport.

Reviewed-by: Daniel Wagner <dwagner@suse.de>
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>