nvme/039: check for logical block size of test device
This test has hard-coded 512 byte values for the dd commands, which
will fail on 4096-byte block devices.  Create an LB_SZ global
variable that is populated with the result of the command
"blockdev --getss <dev>", and use that result for the block size
of the write commands.

Also use this variable for the "--data-len" command of the "nvme
admin-passthru" and "nvme io-passthru" tests.  (On a test with a
4096-byte namespace, leaving these with the hardcoded values still
passed, but update them for the sake of consistency.)

Signed-off-by: Bryan Gurney <bgurney@redhat.com>
Reviewed-by: Alan Adamson <alan.adamson@oracle.com>
Tested-by: Alan Adamson <alan.adamson@oracle.com>
Reviewed-by: Chaitanya Kulkarni <kch@nvidia.com>
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>